- Add MacOS support (gnu-sed is needed. Install with 'brew install gnu-sed') - Fix warnings when creating docker network for the first time - Use litecoin 0.14 - Keep litecoin master branch in cache to avoid having to clone it every time.

+++ altcoin_generator.sh
@@ -44,12 +44,12 @@ LITECOIN_MERKLE_HASH=97ddfbbae6be97fd6cdf3e7ca13232a3afff2353e29badfab7f73011edd
 LITECOIN_MAIN_GENESIS_HASH=12a765e31ffd4059bada1e25190f6e98c99d9714d334efa41a195a7e7e04bfe2
 LITECOIN_TEST_GENESIS_HASH=4966625a4b2851d9fdee139e56211a0d88575f59ed816ff5e6a63deb4e3e29a0
 LITECOIN_REGTEST_GENESIS_HASH=530827f38f93b43ed12af0b3ad25a288dc02ed74d6d7857862df51fc56c416f9
-COIN_NAME_LOWER=${COIN_NAME,,}
-COIN_NAME_UPPER=${COIN_NAME^^}
+COIN_NAME_LOWER=$(echo $COIN_NAME | tr '[:upper:]' '[:lower:]')
+COIN_NAME_UPPER=$(echo $COIN_NAME | tr '[:lower:]' '[:upper:]')
 DIRNAME=$(dirname $0)
 DOCKER_NETWORK="172.18.0"
 DOCKER_IMAGE_LABEL="newcoin-env"
-
+OSVERSION="$(uname -s)"
 
 docker_build_image()
 {
@@ -87,7 +87,7 @@ docker_run()
 
 docker_stop_nodes()
 {
-    echo "Stopping and docker nodes"
+    echo "Stopping all docker nodes"
     for id in $(docker ps -q -a  -f ancestor=$DOCKER_IMAGE_LABEL); do
         docker stop $id
     done
@@ -99,6 +99,14 @@ docker_remove_nodes()
     for id in $(docker ps -q -a  -f ancestor=$DOCKER_IMAGE_LABEL); do
         docker rm $id
     done
+}
+
+docker_create_network()
+{
+    echo "Creating docker network"
+    if ! docker network inspect newcoin &>/dev/null; then
+        docker network create --subnet=$DOCKER_NETWORK.0/16 newcoin
+    fi
 }
 
 docker_remove_network()
@@ -115,11 +123,8 @@ docker_run_node()
     if [ ! -f $DIRNAME/miner${NODE_NUMBER}/$COIN_NAME_LOWER.conf ]; then
         cat <<EOF > $DIRNAME/miner${NODE_NUMBER}/$COIN_NAME_LOWER.conf
 rpcuser=${COIN_NAME_LOWER}rpc
-rpcpassword=$(</dev/urandom tr -dc '12345qwertQWERTasdfgASDFGzxcvbZXCVB' | head -c32)
+rpcpassword=$(cat /dev/urandom | env LC_CTYPE=C tr -dc a-zA-Z0-9 | head -c 32; echo)
 EOF
-    fi
-    if ! docker network inspect newcoin &>/dev/null; then
-        docker network create --subnet=$DOCKER_NETWORK.0/16 newcoin
     fi
 
     docker run --net newcoin --ip $DOCKER_NETWORK.${NODE_NUMBER} -v $DIRNAME/miner${NODE_NUMBER}:/root/.$COIN_NAME_LOWER -v $DIRNAME/$COIN_NAME_LOWER:/$COIN_NAME_LOWER $DOCKER_IMAGE_LABEL /bin/bash -c "$NODE_COMMAND"
@@ -159,18 +164,18 @@ generate_genesis_block()
         cat ${COIN_NAME}-regtest.txt
     fi
 
-    MAIN_PUB_KEY=$(cat ${COIN_NAME}-main.txt | grep "^pubkey:" | sed 's/^pubkey: //')
-    MERKLE_HASH=$(cat ${COIN_NAME}-main.txt | grep "^merkle hash:" | sed 's/^merkle hash: //')
-    TIMESTAMP=$(cat ${COIN_NAME}-main.txt | grep "^time:" | sed 's/^time: //')
-    BITS=$(cat ${COIN_NAME}-main.txt | grep "^bits:" | sed 's/^bits: //')
+    MAIN_PUB_KEY=$(cat ${COIN_NAME}-main.txt | grep "^pubkey:" | $SED 's/^pubkey: //')
+    MERKLE_HASH=$(cat ${COIN_NAME}-main.txt | grep "^merkle hash:" | $SED 's/^merkle hash: //')
+    TIMESTAMP=$(cat ${COIN_NAME}-main.txt | grep "^time:" | $SED 's/^time: //')
+    BITS=$(cat ${COIN_NAME}-main.txt | grep "^bits:" | $SED 's/^bits: //')
 
-    MAIN_NONCE=$(cat ${COIN_NAME}-main.txt | grep "^nonce:" | sed 's/^nonce: //')
-    TEST_NONCE=$(cat ${COIN_NAME}-test.txt | grep "^nonce:" | sed 's/^nonce: //')
-    REGTEST_NONCE=$(cat ${COIN_NAME}-regtest.txt | grep "^nonce:" | sed 's/^nonce: //')
+    MAIN_NONCE=$(cat ${COIN_NAME}-main.txt | grep "^nonce:" | $SED 's/^nonce: //')
+    TEST_NONCE=$(cat ${COIN_NAME}-test.txt | grep "^nonce:" | $SED 's/^nonce: //')
+    REGTEST_NONCE=$(cat ${COIN_NAME}-regtest.txt | grep "^nonce:" | $SED 's/^nonce: //')
 
-    MAIN_GENESIS_HASH=$(cat ${COIN_NAME}-main.txt | grep "^genesis hash:" | sed 's/^genesis hash: //')
-    TEST_GENESIS_HASH=$(cat ${COIN_NAME}-test.txt | grep "^genesis hash:" | sed 's/^genesis hash: //')
-    REGTEST_GENESIS_HASH=$(cat ${COIN_NAME}-regtest.txt | grep "^genesis hash:" | sed 's/^genesis hash: //')
+    MAIN_GENESIS_HASH=$(cat ${COIN_NAME}-main.txt | grep "^genesis hash:" | $SED 's/^genesis hash: //')
+    TEST_GENESIS_HASH=$(cat ${COIN_NAME}-test.txt | grep "^genesis hash:" | $SED 's/^genesis hash: //')
+    REGTEST_GENESIS_HASH=$(cat ${COIN_NAME}-regtest.txt | grep "^genesis hash:" | $SED 's/^genesis hash: //')
 
     popd
 }
@@ -181,61 +186,65 @@ newcoin_replace_vars()
         echo "Warning: $COIN_NAME_LOWER already existing. Not replacing any values"
         return 0
     fi
-    # clone litecoin
-    git clone -b $LITECOIN_BRANCH $LITECOIN_REPOS $COIN_NAME_LOWER
+    if [ ! -d "litecoin-master" ]; then
+        # clone litecoin and keep local cache
+        git clone -b $LITECOIN_BRANCH $LITECOIN_REPOS litecoin-master
+    fi
+
+    git clone -b $LITECOIN_BRANCH litecoin-master $COIN_NAME_LOWER
 
     pushd $COIN_NAME_LOWER
 
     # first rename all directories
     for i in $(find . -type d | grep -v "^./.git" | grep litecoin); do 
-        git mv $i $(echo $i| sed "s/litecoin/$COIN_NAME_LOWER/") 
+        git mv $i $(echo $i| $SED "s/litecoin/$COIN_NAME_LOWER/")
     done
 
     # then rename all files
     for i in $(find . -type f | grep -v "^./.git" | grep litecoin); do
-        git mv $i $(echo $i| sed "s/litecoin/$COIN_NAME_LOWER/")
+        git mv $i $(echo $i| $SED "s/litecoin/$COIN_NAME_LOWER/")
     done
 
     # now replace all litecoin references to the new coin name
     for i in $(find . -type f | grep -v "^./.git"); do
-        sed -i "s/Litecoin/$COIN_NAME/g" $i
-        sed -i "s/litecoin/$COIN_NAME_LOWER/g" $i
-        sed -i "s/LITECOIN/$COIN_NAME_UPPER/g" $i
-        sed -i "s/LTC/$COIN_UNIT/g" $i
+        $SED -i "s/Litecoin/$COIN_NAME/g" $i
+        $SED -i "s/litecoin/$COIN_NAME_LOWER/g" $i
+        $SED -i "s/LITECOIN/$COIN_NAME_UPPER/g" $i
+        $SED -i "s/LTC/$COIN_UNIT/g" $i
     done
 
-    sed -i "s/84000000/$TOTAL_SUPPLY/" src/amount.h
-    sed -i "s/1,48/1,$PUBKEY_CHAR/" src/chainparams.cpp
+    $SED -i "s/84000000/$TOTAL_SUPPLY/" src/amount.h
+    $SED -i "s/1,48/1,$PUBKEY_CHAR/" src/chainparams.cpp
 
-    sed -i "s/1317972665/$TIMESTAMP/" src/chainparams.cpp
+    $SED -i "s/1317972665/$TIMESTAMP/" src/chainparams.cpp
 
-    sed -i "s;NY Times 05/Oct/2011 Steve Jobs, Apple’s Visionary, Dies at 56;$PHRASE;" src/chainparams.cpp
+    $SED -i "s;NY Times 05/Oct/2011 Steve Jobs, Apple’s Visionary, Dies at 56;$PHRASE;" src/chainparams.cpp
 
-    sed -i "s/= 9333;/= $MAINNET_PORT;/" src/chainparams.cpp
-    sed -i "s/= 19335;/= $TESTNET_PORT;/" src/chainparams.cpp
+    $SED -i "s/= 9333;/= $MAINNET_PORT;/" src/chainparams.cpp
+    $SED -i "s/= 19335;/= $TESTNET_PORT;/" src/chainparams.cpp
 
-    sed -i "s/$LITECOIN_PUB_KEY/$MAIN_PUB_KEY/" src/chainparams.cpp
-    sed -i "s/$LITECOIN_MERKLE_HASH/$MERKLE_HASH/" src/chainparams.cpp
-    sed -i "s/$LITECOIN_MERKLE_HASH/$MERKLE_HASH/" src/qt/test/rpcnestedtests.cpp
+    $SED -i "s/$LITECOIN_PUB_KEY/$MAIN_PUB_KEY/" src/chainparams.cpp
+    $SED -i "s/$LITECOIN_MERKLE_HASH/$MERKLE_HASH/" src/chainparams.cpp
+    $SED -i "s/$LITECOIN_MERKLE_HASH/$MERKLE_HASH/" src/qt/test/rpcnestedtests.cpp
 
-    sed -i "0,/$LITECOIN_MAIN_GENESIS_HASH/s//$MAIN_GENESIS_HASH/" src/chainparams.cpp
-    sed -i "0,/$LITECOIN_TEST_GENESIS_HASH/s//$TEST_GENESIS_HASH/" src/chainparams.cpp
-    sed -i "0,/$LITECOIN_REGTEST_GENESIS_HASH/s//$REGTEST_GENESIS_HASH/" src/chainparams.cpp
+    $SED -i "0,/$LITECOIN_MAIN_GENESIS_HASH/s//$MAIN_GENESIS_HASH/" src/chainparams.cpp
+    $SED -i "0,/$LITECOIN_TEST_GENESIS_HASH/s//$TEST_GENESIS_HASH/" src/chainparams.cpp
+    $SED -i "0,/$LITECOIN_REGTEST_GENESIS_HASH/s//$REGTEST_GENESIS_HASH/" src/chainparams.cpp
 
-    sed -i "0,/2084524493/s//$MAIN_NONCE/" src/chainparams.cpp
-    sed -i "0,/293345/s//$TEST_NONCE/" src/chainparams.cpp
-    sed -i "0,/1296688602, 0/s//1296688602, $REGTEST_NONCE/" src/chainparams.cpp
-    sed -i "0,/0x1e0ffff0/s//$BITS/" src/chainparams.cpp
+    $SED -i "0,/2084524493/s//$MAIN_NONCE/" src/chainparams.cpp
+    $SED -i "0,/293345/s//$TEST_NONCE/" src/chainparams.cpp
+    $SED -i "0,/1296688602, 0/s//1296688602, $REGTEST_NONCE/" src/chainparams.cpp
+    $SED -i "0,/0x1e0ffff0/s//$BITS/" src/chainparams.cpp
 
-    sed -i "s,vSeeds.push_back,//vSeeds.push_back,g" src/chainparams.cpp
+    $SED -i "s,vSeeds.push_back,//vSeeds.push_back,g" src/chainparams.cpp
 
     if [ -n "$PREMINED_AMOUNT" ]; then
-        sed -i "s/CAmount nSubsidy = 50 \* COIN;/if \(nHeight == 1\) return COIN \* $PREMINED_AMOUNT;\n    CAmount nSubsidy = 50 \* COIN;/" src/validation.cpp
+        $SED -i "s/CAmount nSubsidy = 50 \* COIN;/if \(nHeight == 1\) return COIN \* $PREMINED_AMOUNT;\n    CAmount nSubsidy = 50 \* COIN;/" src/validation.cpp
     fi
 
-    sed -i "s/COINBASE_MATURITY = 100/COINBASE_MATURITY = $COINBASE_MATURITY/" src/consensus/consensus.h
+    $SED -i "s/COINBASE_MATURITY = 100/COINBASE_MATURITY = $COINBASE_MATURITY/" src/consensus/consensus.h
 
-    sed -i "s/0x00000000000000000000000000000000000000000000000ba50a60f8b56c7fe0/$MINIMUM_CHAIN_WORK/" src/chainparams.cpp
+    $SED -i "s/0x00000000000000000000000000000000000000000000000ba50a60f8b56c7fe0/$MINIMUM_CHAIN_WORK/" src/chainparams.cpp
 
     # TODO: fix checkpoints
     popd
@@ -260,6 +269,26 @@ fi
 cd $DIRNAME
 
 # sanity check
+
+case $OSVERSION in
+    Linux*)
+        SED=sed
+    ;;
+    Darwin*)
+        SED=$(which gsed 2>/dev/null)
+        if [ $? -ne 0 ]; then
+            echo "please install gnu-sed with 'brew install gnu-sed'"
+            exit 1
+        fi
+        SED=gsed
+    ;;
+    *)
+        echo "This script only works on Linux and MacOS"
+        exit 1
+    ;;
+esac
+
+
 if ! which docker &>/dev/null; then
     echo Please install docker first
     exit 1
@@ -275,22 +304,23 @@ case $1 in
         docker_stop_nodes
     ;;
     remove_nodes)
-	docker_stop_nodes
-	docker_remove_nodes
+        docker_stop_nodes
+        docker_remove_nodes
     ;;
     clean_up)
         docker_stop_nodes
         for i in $(seq 2 5); do
            docker_run_node $i "rm -rf /$COIN_NAME_LOWER /root/.$COIN_NAME_LOWER" &>/dev/null
         done
-	docker_remove_nodes
+        docker_remove_nodes
         docker_remove_network
-	rm -rf $COIN_NAME_LOWER
-	rm -f GenesisH0/${COIN_NAME}-*.txt
+        rm -rf $COIN_NAME_LOWER
+        if [ "$2" != "keep_genesis_block" ]; then
+            rm -f GenesisH0/${COIN_NAME}-*.txt
+        fi
         for i in $(seq 2 5); do
            rm -rf miner$i
         done
-
     ;;
     start)
         if [ -n "$(docker ps -q -f ancestor=$DOCKER_IMAGE_LABEL)" ]; then
@@ -301,6 +331,7 @@ case $1 in
         generate_genesis_block
         newcoin_replace_vars
         build_new_coin
+        docker_create_network
 
         docker_run_node 2 "cd /$COIN_NAME_LOWER ; ./src/${COIN_NAME_LOWER}d $CHAIN -listen -noconnect -bind=$DOCKER_NETWORK.2 -addnode=$DOCKER_NETWORK.1 -addnode=$DOCKER_NETWORK.3 -addnode=$DOCKER_NETWORK.4 -addnode=$DOCKER_NETWORK.5" &
         docker_run_node 3 "cd /$COIN_NAME_LOWER ; ./src/${COIN_NAME_LOWER}d $CHAIN -listen -noconnect -bind=$DOCKER_NETWORK.3 -addnode=$DOCKER_NETWORK.1 -addnode=$DOCKER_NETWORK.2 -addnode=$DOCKER_NETWORK.4 -addnode=$DOCKER_NETWORK.5" &
@@ -321,4 +352,5 @@ Usage: $0 (start|stop|remove_nodes|clean_up)
  - remove_nodes: remove the old docker container images. This will stop them first if necessary.
  - clean_up: WARNING: this will stop and remove docker containers and network, source code, genesis block information and nodes data directory. (to start from scratch)
 EOF
+    ;;
 esac
